+++ nieumarli/src/test/java/com/s14222/tau/UndeadTest.java
@@ -24,7 +24,7 @@ public class UndeadTest {
 
     @Test
     public void TestDodania(){
-        Undead undead1 = new Undead(7, "Trups",1,100,100);
+        Undead undead1 = new Undead((long) 7, "Trups",1,100,100);
 
         undeadRepository.add(undead1);
 
@@ -35,11 +35,11 @@ public class UndeadTest {
     @Test
     public void TestUsuniecia() {
 
-        undeadRepository.deleteById(6);
+        undeadRepository.deleteById((long) 6);
 
         List<Undead> undeads = undeadRepository.getAll();
 
-        assertNull(undeadRepository.getById(6).getNazwa());
+        assertNull(undeadRepository.getById((long) 6).getNazwa());
         assertEquals(true, !undeads.isEmpty());
 
         }
@@ -47,16 +47,16 @@ public class UndeadTest {
     @Test
     public void TestUpdate() {
 
-        Undead undeadToUpdate = undeadRepository.getById(2);
+        Undead undeadToUpdate = undeadRepository.getById((long)2);
 
         undeadToUpdate.setNazwa("Zombie");
 
         undeadRepository.updateById(undeadToUpdate);
 
-        assertThat(undeadRepository.getById(2).getNazwa(), is(undeadToUpdate.getNazwa()));
-        assertEquals(undeadRepository.getById(2).getNazwa(), undeadToUpdate.getNazwa());
+        assertThat(undeadRepository.getById((long) 2).getNazwa(), is(undeadToUpdate.getNazwa()));
+        assertEquals(undeadRepository.getById((long) 2).getNazwa(), undeadToUpdate.getNazwa());
 
-        assertFalse("nie powinno modyfikowac",undeadRepository.getById(4).getNazwa().equals(undeadToUpdate.getNazwa()));
+        assertFalse("nie powinno modyfikowac",undeadRepository.getById((long)4).getNazwa().equals(undeadToUpdate.getNazwa()));
 
 
     }
@@ -84,12 +84,12 @@ public class UndeadTest {
     public void initRepository(){
         undeadRepository = UndeadRepositoryFactory.getInstance();
 
-        Undead zeroUndead = new Undead(0, "Trups", 1, 100, 100);
-        Undead firstUndead = new Undead(1 , "Duch", 1, 200, 200);
-        Undead secondUndead = new Undead(2, "Wampir", 2,200,150);
-        Undead thirdUndead = new Undead(4, "Wilkolak", 1,100,100);
-        Undead fourthUndead = new Undead(5, "Gnom", 4, 45, 50);
-        Undead fifthUndead = new Undead(6, "Elf", 2, 120, 1);
+        Undead zeroUndead = new Undead((long) 0, "Trups", 1, 100, 100);
+        Undead firstUndead = new Undead((long) 1, "Duch", 1, 200, 200);
+        Undead secondUndead = new Undead((long) 2, "Wampir", 2,200,150);
+        Undead thirdUndead = new Undead((long) 4, "Wilkolak", 1,100,100);
+        Undead fourthUndead = new Undead((long) 5, "Gnom", 4, 45, 50);
+        Undead fifthUndead = new Undead((long) 6, "Elf", 2, 120, 1);
 
         undeadRepository.add(zeroUndead);
         undeadRepository.add(firstUndead);
